refactor: admin 위치 변경
(#273)
